show_common: add file for *_show common functions
Move functions being used by multiple *_show commands to a common file,
providing a better code organization and thus, easing developers' lives.
